refactor(engine): extract fsm/session/photo/response modules from engine.py (#527)
Extract four independently testable modules from the 2930-line engine.py monolith:
- fsm.py — STATE_ORDER, VALID_STATES, _STATE_ALIASES, advance_state()
- session_manager.py — SQLite load/save/record/log helpers
- photo_handler.py — session photo I/O + build_print_reply()
- response_formatter.py — parse_response(), format_reply(), regex constants

engine.py becomes a thin orchestrator delegating to these modules.
Backward-compat re-exports preserved for tests that import from shared.engine.
Also restores missing _build_wo_draft() method and updates test_source_file_consistency
to check fsm.py (where aliases now live).
